Upgraded kaiko to v3 framework
