+++ .changeset/few-poems-judge.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/kaiko-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
