--- packages/sources/kaiko-test/CHANGELOG.md
@@ -1,126 +1,0 @@
-# @chainlink/kaiko-test-adapter
-
-## 1.1.15
-
-### Patch Changes
-
-- f8d90dc59: Added 60s API timeout override
-
-## 1.1.14
-
-### Patch Changes
-
-- 1e8adb216: Fixed response parsing to skip null values
-
-## 1.1.13
-
-### Patch Changes
-
-- cfac404db: Added 0 price validation
-
-## 1.1.12
-
-### Patch Changes
-
-- 41f253afb: Fixed millisecondsAgo input param type
-
-## 1.1.11
-
-### Patch Changes
-
-- 6cdb337f2: Updated rate limit tier name
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.1.10
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.1.9
-
-### Patch Changes
-
-- 29dd7b460: Set default override for MAX_HTTP_REQUEST_QUEUE_LENGTH to 300
-- 31af84e69: Bumped framework version
-
-## 1.1.8
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.1.7
-
-### Patch Changes
-
-- 72296e826: Added reasonable rate limit tier
-
-## 1.1.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.1.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.1.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.3
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.0.2
-
-### Patch Changes
-
-- c152ed06e: Bumped v3 framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- fa05cc85b: Removed rate limit config
-
-## 1.0.0
-
-### Major Changes
-
-- bba54b419: Added kaiko-test EA
