--- packages/sources/kaiko-test/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@chainlink/kaiko-test-adapter",
-  "version": "1.1.15",
-  "description": "Chainlink kaiko-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "kaiko-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.29.11",
-    "tslib": "^2.3.1"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "@types/supertest": "2.0.12",
-    "nock": "13.2.9",
-    "supertest": "6.2.4",
-    "typescript": "5.0.4"
-  }
-}
