--- packages/sources/kaiko-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as trades } from './trades'
