--- packages/sources/kaiko-test/test/integration/adapter.test.ts
@@ -1,79 +1,0 @@
-import { SuperTest, Test } from 'supertest'
-import { setupExternalAdapterTest, SuiteContext } from './setup'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
-import { mockRateResponseFailure, mockRateResponseSuccess } from './fixtures'
-
-describe('execute', () => {
-  const id = '1'
-  let spy: jest.SpyInstance
-  beforeAll(async () => {
-    const mockDate = new Date('2022-01-01T11:11:11.111Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-  })
-
-  afterAll((done) => {
-    spy.mockRestore()
-    done()
-  })
-
-  const context: SuiteContext = {
-    req: null,
-    server: async () => {
-      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '6'
-      process.env['METRICS_ENABLED'] = 'false'
-      process.env['API_KEY'] = 'fake-api-key'
-      const server = (await import('../../src')).server
-      return server() as Promise<ServerInstance>
-    },
-  }
-
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('trades endpoint', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockRateResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-
-    xit('should return failure', async () => {
-      mockRateResponseFailure()
-
-      const data = {
-        id,
-        data: {
-          base: 'XXX',
-          quote: 'USD',
-        },
-      }
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-})
