+++ packages/sources/kaiko/package.json
@@ -28,8 +28,7 @@
     "start": "yarn server:dist"
   },
   "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/ea-test-helpers": "workspace:*",
+    "@chainlink/external-adapter-framework": "0.29.11",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
