--- packages/sources/kaiko/schemas/env.json
@@ -1,20 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/kaiko-adapter.json",
-  "title": "Chainlink External Adapter for Kaiko",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string"
-    },
-    "API_ENDPOINT": {
-      "default": "https://us.market-api.kaiko.io/v2/data/trades.v1",
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
