--- packages/sources/kaiko/src/config/symbols.json
@@ -1,5 +1,0 @@
-{
-  "kaiko": {
-    "RAI": "RAIR"
-  }
-}
