+++ packages/sources/kaiko/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as TradesInputParameters } from './trades'
-
-export type TInputParameters = TradesInputParameters
-
-export * as trades from './trades'
+export { endpoint as trades } from './trades'
