--- packages/sources/kaiko/test/e2e/adapter.test.ts
@@ -1,94 +1,0 @@
-import { AdapterRequest, FastifyInstance } from '@chainlink/ea-bootstrap'
-import request, { SuperTest, Test } from 'supertest'
-import * as process from 'process'
-import { server as startServer } from '../../src'
-import { AddressInfo } from 'net'
-
-/* Since no API keys are set, all requests will error, but the error will
-  confirm that the correct endpoint is being used. */
-
-describe('execute', () => {
-  const id = '1'
-  let fastify: FastifyInstance
-  let req: SuperTest<Test>
-
-  beforeAll(async () => {
-    process.env.CACHE_ENABLED = 'false'
-    process.env.API_KEY = 'fake-api-key'
-    fastify = await startServer()
-    req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
-  })
-
-  afterAll((done) => {
-    fastify.close(done)
-  })
-
-  describe('exchange rate api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      const response = await req
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body.error.url).toBe(
-        'https:/us.market-api.kaiko.io/v2/data/trades.v1/spot_exchange_rate/eth/usd',
-      )
-    })
-  })
-
-  describe('exchange direct rate api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'LTC',
-        quote: 'ETH',
-      },
-    }
-
-    it('should return success', async () => {
-      const response = await req
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body.error.url).toBe(
-        'https:/us.market-api.kaiko.io/v2/data/trades.v1/spot_exchange_rate/ltc/eth',
-      )
-    })
-  })
-
-  describe('exchange rate api with invalid token', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'XXX',
-        quote: 'EUR',
-      },
-    }
-
-    it('should return failure', async () => {
-      const response = await req
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body.error.url).toBe(
-        'https:/us.market-api.kaiko.io/v2/data/trades.v1/spot_exchange_rate/xxx/eur',
-      )
-    })
-  })
-})
