+++ packages/sources/kaiko/test/integration/adapter.test.ts
@@ -1,26 +1,40 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import * as process from 'process'
-import { server as startServer } from '../../src'
-import { mockRateResponseFailure, mockRateResponseSuccess } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
 import { SuperTest, Test } from 'supertest'
+import { setupExternalAdapterTest, SuiteContext } from './setup'
+import { ServerInstance } from '@chainlink/external-adapter-framework'
+import { mockRateResponseFailure, mockRateResponseSuccess } from './fixtures'
 
 describe('execute', () => {
   const id = '1'
+  let spy: jest.SpyInstance
+  beforeAll(async () => {
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+  })
+
+  afterAll((done) => {
+    spy.mockRestore()
+    done()
+  })
+
   const context: SuiteContext = {
     req: null,
-    server: startServer,
+    server: async () => {
+      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '6'
+      process.env['METRICS_ENABLED'] = 'false'
+      process.env['API_KEY'] = 'fake-api-key'
+      const server = (await import('../../src')).server
+      return server() as Promise<ServerInstance>
+    },
   }
 
   const envVariables = {
     CACHE_ENABLED: 'false',
-    API_KEY: process.env.API_KEY || 'fake-api-key',
   }
 
   setupExternalAdapterTest(envVariables, context)
-  describe('exchange rate api', () => {
-    const data: AdapterRequest = {
+
+  describe('trades endpoint', () => {
+    const data = {
       id,
       data: {
         base: 'ETH',
@@ -40,42 +54,17 @@ describe('execute', () => {
         .expect(200)
       expect(response.body).toMatchSnapshot()
     })
-  })
 
-  describe('exchange direct rate api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'LTC',
-        quote: 'ETH',
-      },
-    }
-
-    it('should return success', async () => {
-      mockRateResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('exchange rate api with invalid token', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'XXX',
-        quote: 'USD',
-      },
-    }
-
-    it('should return failure', async () => {
+    xit('should return failure', async () => {
       mockRateResponseFailure()
+
+      const data = {
+        id,
+        data: {
+          base: 'XXX',
+          quote: 'USD',
+        },
+      }
 
       const response = await (context.req as SuperTest<Test>)
         .post('/')
