+++ packages/sources/kaiko/tsconfig.json
@@ -5,6 +5,5 @@
     "rootDir": "src"
   },
   "include": ["src/**/*", "src/**/*.json"],
-  "exclude": ["dist", "**/*.spec.ts", "**/*.test.ts"],
-  "references": [{ "path": "../../core/test-helpers" }, { "path": "../../core/bootstrap" }]
+  "exclude": ["dist", "**/*.spec.ts", "**/*.test.ts"]
 }
