+++ packages/tsconfig.json
@@ -396,9 +396,6 @@
       "path": "./sources/kaiko"
     },
     {
-      "path": "./sources/kaiko-test"
-    },
-    {
       "path": "./sources/layer2-sequencer-health"
     },
     {
